io: extract io_conn_* bookkeeping into conn_info.c
The io_conn_* connection-tracking module (per-fd op counts, state
transitions, timeout sweeps, write-serialization gate) is backend-
agnostic -- it touches neither io_uring nor kqueue.  Before this
change, it lived in connect.c (Linux-only build target), and the
FreeBSD kqueue backend had ~20 no-op stubs standing in for it in
backend_kqueue.c.

Extract the full ~700-line bookkeeping block (plus io_socket_close,
io_conn_dump*, and the state/role stringifiers) into lib/io/conn_info.c.
Compile it in both backend builds.  Remove the corresponding stubs
from backend_kqueue.c.  connect.c now contains only io_send_request
and io_handle_connect -- the io_uring-specific connect-op logic.

No functional change on Linux.  On FreeBSD, the kqueue backend now
has real connection tracking instead of silently dropping every op
count.  Completion handlers (io_handle_accept/connect/read/write) and
buffer-state management remain stubs; those are later PRs.
