+++ lib/io/connect.c
@@ -1,6 +1,11 @@
 /*
  * SPDX-FileCopyrightText: 2025 Tom Haynes <loghyr@gmail.com>
  * SPDX-License-Identifier: AGPL-3.0-or-later
+ *
+ * io_uring connect-operation path: socket creation, io_uring_prep_connect
+ * submission, and the io_handle_connect completion handler.  Shared
+ * connection bookkeeping (io_conn_register/get/add_*_op/...) lives in
+ * conn_info.c.
  */
 
 #ifdef HAVE_CONFIG_H
@@ -12,7 +17,6 @@
 #include <fcntl.h>
 #include <liburing.h>
 #include <netinet/in.h>
-#include <openssl/ssl.h>
 #include <pthread.h>
 #include <stdbool.h>
 #include <stdint.h>
@@ -24,774 +28,11 @@
 
 #include "reffs/log.h"
 #include "reffs/io.h"
-#include "reffs/log.h"
 #include "reffs/network.h"
 #include "reffs/ring.h"
 #include "ring_internal.h"
 #include "reffs/rpc.h"
 #include "reffs/trace/io.h"
-
-// Array to track connection states
-static struct conn_info *connections[MAX_CONNECTIONS];
-static pthread_mutex_t conn_mutex = PTHREAD_MUTEX_INITIALIZER;
-
-// Initialize connection tracking
-int io_conn_init(void)
-{
-	pthread_mutex_lock(&conn_mutex);
-	memset(connections, 0, sizeof(connections));
-	pthread_mutex_unlock(&conn_mutex);
-	return 0;
-}
-
-// Register a new connection
-struct conn_info *io_conn_register(int fd, enum conn_state initial_state,
-				   enum conn_role role)
-{
-	struct conn_info *ci = NULL;
-
-	pthread_mutex_lock(&conn_mutex);
-
-	// Find unused slot or reuse existing slot for this fd
-	int idx = fd % MAX_CONNECTIONS;
-	if (connections[idx] && connections[idx]->ci_fd == fd) {
-		ci = connections[idx];
-	} else if (!connections[idx] ||
-		   connections[idx]->ci_state == CONN_UNUSED) {
-		// Allocate new conn_info if needed
-		if (!connections[idx]) {
-			connections[idx] = malloc(sizeof(struct conn_info));
-			if (!connections[idx]) {
-				pthread_mutex_unlock(&conn_mutex);
-				return NULL;
-			}
-		}
-
-		ci = connections[idx];
-		memset(ci, 0, sizeof(struct conn_info));
-		ci->ci_fd = fd;
-	} else {
-		// Collision with another active connection
-		LOG("Connection slot collision for fd=%d", fd);
-	}
-
-	if (ci) {
-		ci->ci_state = initial_state;
-		ci->ci_role = role;
-		ci->ci_last_activity = time(NULL);
-
-		/* Initialize operation counters */
-		ci->ci_read_count = 0;
-		ci->ci_write_count = 0;
-		ci->ci_accept_count = 0;
-		ci->ci_connect_count = 0;
-
-		/* Initialize per-fd write serialization gate */
-		ci->ci_write_active = false;
-		ci->ci_write_pending_head = NULL;
-		ci->ci_write_pending_tail = NULL;
-	}
-
-	pthread_mutex_unlock(&conn_mutex);
-	return ci;
-}
-
-// Get connection info
-struct conn_info *io_conn_get(int fd)
-{
-	struct conn_info *ci = NULL;
-	pthread_mutex_lock(&conn_mutex);
-
-	int idx = fd % MAX_CONNECTIONS;
-	if (connections[idx] && connections[idx]->ci_fd == fd) {
-		ci = connections[idx];
-	}
-
-	pthread_mutex_unlock(&conn_mutex);
-	return ci;
-}
-
-// Update the connection state based on operation counts
-void io_conn_update_state(int fd)
-{
-	pthread_mutex_lock(&conn_mutex);
-
-	int idx = fd % MAX_CONNECTIONS;
-	if (connections[idx] && connections[idx]->ci_fd == fd) {
-		enum conn_state old_state = connections[idx]->ci_state;
-		enum conn_state new_state;
-
-		// Determine the new state based on operation counts
-		if (connections[idx]->ci_read_count > 0) {
-			if (connections[idx]->ci_write_count > 0) {
-				new_state = CONN_READWRITE;
-			} else {
-				new_state = CONN_READING;
-			}
-		} else if (connections[idx]->ci_write_count > 0) {
-			new_state = CONN_WRITING;
-		} else if (connections[idx]->ci_accept_count > 0) {
-			new_state = CONN_ACCEPTING;
-		} else if (connections[idx]->ci_connect_count > 0) {
-			new_state = CONN_CONNECTING;
-		} else if (old_state == CONN_ERROR ||
-			   old_state == CONN_DISCONNECTING ||
-			   old_state == CONN_UNUSED) {
-			// Don't change these special states if no ops are pending
-			new_state = old_state;
-		} else {
-			new_state = CONN_CONNECTED;
-		}
-
-		// Only log if state actually changes
-		if (old_state != new_state) {
-			trace_io_connection_state_change(
-				fd, old_state, new_state, __func__, __LINE__);
-			connections[idx]->ci_state = new_state;
-		}
-	}
-
-	pthread_mutex_unlock(&conn_mutex);
-}
-
-// Function to add a read operation
-int io_conn_add_read_op(int fd)
-{
-	int ret = -1;
-	pthread_mutex_lock(&conn_mutex);
-
-	int idx = fd % MAX_CONNECTIONS;
-	if (connections[idx] && connections[idx]->ci_fd == fd) {
-		connections[idx]->ci_read_count++;
-		connections[idx]->ci_last_activity = time(NULL);
-
-		// Update the connection state based on the new count
-		enum conn_state old_state = connections[idx]->ci_state;
-		enum conn_state new_state;
-
-		if (connections[idx]->ci_write_count > 0) {
-			new_state = CONN_READWRITE;
-		} else {
-			new_state = CONN_READING;
-		}
-
-		if (old_state != new_state) {
-			trace_io_connection_state_change(
-				fd, old_state, new_state, __func__, __LINE__);
-			connections[idx]->ci_state = new_state;
-		}
-
-		ret = 0;
-	}
-
-	pthread_mutex_unlock(&conn_mutex);
-	return ret;
-}
-
-// Function to remove a read operation
-int io_conn_remove_read_op(int fd)
-{
-	int ret = -1;
-	pthread_mutex_lock(&conn_mutex);
-
-	int idx = fd % MAX_CONNECTIONS;
-	if (connections[idx] && connections[idx]->ci_fd == fd) {
-		if (connections[idx]->ci_read_count > 0) {
-			connections[idx]->ci_read_count--;
-			trace_io_connection_count(
-				fd, connections[idx]->ci_read_count, __func__,
-				__LINE__);
-
-			// Update state based on remaining operations
-			enum conn_state old_state = connections[idx]->ci_state;
-			enum conn_state new_state;
-
-			if (connections[idx]->ci_read_count > 0) {
-				if (connections[idx]->ci_write_count > 0) {
-					new_state = CONN_READWRITE;
-				} else {
-					new_state = CONN_READING;
-				}
-			} else if (connections[idx]->ci_write_count > 0) {
-				new_state = CONN_WRITING;
-			} else {
-				new_state = CONN_CONNECTED;
-			}
-
-			if (old_state != new_state) {
-				trace_io_connection_state_change(fd, old_state,
-								 new_state,
-								 __func__,
-								 __LINE__);
-				connections[idx]->ci_state = new_state;
-			}
-		} else {
-			LOG("Warning: Attempt to remove read op when count is zero for fd=%d",
-			    fd);
-		}
-		connections[idx]->ci_last_activity = time(NULL);
-		ret = 0;
-	}
-
-	pthread_mutex_unlock(&conn_mutex);
-	return ret;
-}
-
-// Function to add a write operation
-int io_conn_add_write_op(int fd)
-{
-	int ret = -1;
-	pthread_mutex_lock(&conn_mutex);
-
-	int idx = fd % MAX_CONNECTIONS;
-	if (connections[idx] && connections[idx]->ci_fd == fd) {
-		connections[idx]->ci_write_count++;
-		connections[idx]->ci_last_activity = time(NULL);
-		trace_io_connection_count(fd, connections[idx]->ci_write_count,
-					  __func__, __LINE__);
-
-		// Update the connection state based on the new count
-		enum conn_state old_state = connections[idx]->ci_state;
-		enum conn_state new_state;
-
-		if (connections[idx]->ci_read_count > 0) {
-			new_state = CONN_READWRITE;
-		} else {
-			new_state = CONN_WRITING;
-		}
-
-		if (old_state != new_state) {
-			trace_io_connection_state_change(
-				fd, old_state, new_state, __func__, __LINE__);
-			connections[idx]->ci_state = new_state;
-		}
-
-		ret = 0;
-	}
-
-	pthread_mutex_unlock(&conn_mutex);
-	return ret;
-}
-
-// Function to remove a write operation
-int io_conn_remove_write_op(int fd)
-{
-	int ret = -1;
-	pthread_mutex_lock(&conn_mutex);
-
-	int idx = fd % MAX_CONNECTIONS;
-	if (connections[idx] && connections[idx]->ci_fd == fd) {
-		if (connections[idx]->ci_write_count > 0) {
-			connections[idx]->ci_write_count--;
-
-			trace_io_connection_count(
-				fd, connections[idx]->ci_write_count, __func__,
-				__LINE__);
-
-			// Update state based on remaining operations
-			enum conn_state old_state = connections[idx]->ci_state;
-			enum conn_state new_state;
-
-			if (connections[idx]->ci_write_count > 0) {
-				if (connections[idx]->ci_read_count > 0) {
-					new_state = CONN_READWRITE;
-				} else {
-					new_state = CONN_WRITING;
-				}
-			} else if (connections[idx]->ci_read_count > 0) {
-				new_state = CONN_READING;
-			} else {
-				new_state = CONN_CONNECTED;
-			}
-
-			if (old_state != new_state) {
-				trace_io_connection_state_change(fd, old_state,
-								 new_state,
-								 __func__,
-								 __LINE__);
-				connections[idx]->ci_state = new_state;
-			}
-		} else {
-			LOG("Warning: Attempt to remove write op when count is zero for fd=%d",
-			    fd);
-		}
-		connections[idx]->ci_last_activity = time(NULL);
-		ret = 0;
-	}
-
-	pthread_mutex_unlock(&conn_mutex);
-	return ret;
-}
-
-// Function to add an accept operation
-int io_conn_add_accept_op(int fd)
-{
-	int ret = -1;
-	pthread_mutex_lock(&conn_mutex);
-
-	int idx = fd % MAX_CONNECTIONS;
-	if (connections[idx] && connections[idx]->ci_fd == fd) {
-		connections[idx]->ci_accept_count++;
-		connections[idx]->ci_last_activity = time(NULL);
-		trace_io_connection_count(fd, connections[idx]->ci_accept_count,
-					  __func__, __LINE__);
-
-		// Update state to ACCEPTING
-		if (connections[idx]->ci_state != CONN_ACCEPTING) {
-			trace_io_connection_state_change(
-				fd, connections[idx]->ci_state, CONN_ACCEPTING,
-				__func__, __LINE__);
-			connections[idx]->ci_state = CONN_ACCEPTING;
-		}
-
-		ret = 0;
-	}
-
-	pthread_mutex_unlock(&conn_mutex);
-	return ret;
-}
-
-// Function to remove an accept operation
-int io_conn_remove_accept_op(int fd)
-{
-	int ret = -1;
-	pthread_mutex_lock(&conn_mutex);
-
-	int idx = fd % MAX_CONNECTIONS;
-	if (connections[idx] && connections[idx]->ci_fd == fd) {
-		if (connections[idx]->ci_accept_count > 0) {
-			connections[idx]->ci_accept_count--;
-			trace_io_connection_count(
-				fd, connections[idx]->ci_accept_count, __func__,
-				__LINE__);
-
-			// Update state based on remaining operations
-			if (connections[idx]->ci_accept_count == 0 &&
-			    connections[idx]->ci_state == CONN_ACCEPTING) {
-				trace_io_connection_state_change(
-					fd, CONN_ACCEPTING, CONN_LISTENING,
-					__func__, __LINE__);
-				connections[idx]->ci_state = CONN_LISTENING;
-			}
-		} else {
-			LOG("Warning: Attempt to remove accept op when count is zero for fd=%d",
-			    fd);
-		}
-		connections[idx]->ci_last_activity = time(NULL);
-		ret = 0;
-	}
-
-	pthread_mutex_unlock(&conn_mutex);
-	return ret;
-}
-
-// Function to add a connect operation
-int io_conn_add_connect_op(int fd)
-{
-	int ret = -1;
-	pthread_mutex_lock(&conn_mutex);
-
-	int idx = fd % MAX_CONNECTIONS;
-	if (connections[idx] && connections[idx]->ci_fd == fd) {
-		connections[idx]->ci_connect_count++;
-		connections[idx]->ci_last_activity = time(NULL);
-		trace_io_connection_count(fd,
-					  connections[idx]->ci_connect_count,
-					  __func__, __LINE__);
-
-		// Update state to CONNECTING
-		if (connections[idx]->ci_state != CONN_CONNECTING) {
-			trace_io_connection_state_change(
-				fd, connections[idx]->ci_state, CONN_CONNECTING,
-				__func__, __LINE__);
-			connections[idx]->ci_state = CONN_CONNECTING;
-		}
-
-		ret = 0;
-	}
-
-	pthread_mutex_unlock(&conn_mutex);
-	return ret;
-}
-
-// Function to remove a connect operation
-int io_conn_remove_connect_op(int fd)
-{
-	int ret = -1;
-
-	pthread_mutex_lock(&conn_mutex);
-
-	int idx = fd % MAX_CONNECTIONS;
-	if (connections[idx] && connections[idx]->ci_fd == fd) {
-		if (connections[idx]->ci_connect_count > 0) {
-			connections[idx]->ci_connect_count--;
-			trace_io_connection_count(
-				fd, connections[idx]->ci_connect_count,
-				__func__, __LINE__);
-
-			// Update state based on remaining operations
-			if (connections[idx]->ci_connect_count == 0 &&
-			    connections[idx]->ci_state == CONN_CONNECTING) {
-				trace_io_connection_state_change(
-					fd, CONN_CONNECTING, CONN_CONNECTED,
-					__func__, __LINE__);
-				connections[idx]->ci_state = CONN_CONNECTED;
-			}
-		} else {
-			LOG("Warning: Attempt to remove connect op when count is zero for fd=%d",
-			    fd);
-		}
-		connections[idx]->ci_last_activity = time(NULL);
-		ret = 0;
-	}
-
-	pthread_mutex_unlock(&conn_mutex);
-	return ret;
-}
-
-// Set connection to error state
-int io_conn_set_error(int fd, int error_code)
-{
-	int ret = -1;
-	pthread_mutex_lock(&conn_mutex);
-
-	int idx = fd % MAX_CONNECTIONS;
-	if (connections[idx] && connections[idx]->ci_fd == fd) {
-		trace_io_connection_state_change(fd, connections[idx]->ci_state,
-						 CONN_ERROR, __func__,
-						 __LINE__);
-
-		connections[idx]->ci_state = CONN_ERROR;
-		connections[idx]->ci_error = error_code;
-		connections[idx]->ci_last_activity = time(NULL);
-		ret = 0;
-	}
-
-	pthread_mutex_unlock(&conn_mutex);
-	return ret;
-}
-
-// Check if a connection is in a particular state
-bool io_conn_is_state(int fd, enum conn_state state)
-{
-	bool result = false;
-	pthread_mutex_lock(&conn_mutex);
-
-	int idx = fd % MAX_CONNECTIONS;
-	if (connections[idx] && connections[idx]->ci_fd == fd) {
-		result = (connections[idx]->ci_state == state);
-	}
-
-	pthread_mutex_unlock(&conn_mutex);
-	return result;
-}
-
-void io_conn_destroy(struct conn_info *ci)
-{
-	if (!ci)
-		return;
-
-	// Clean up SSL if present
-	if (ci->ci_ssl) {
-		SSL_shutdown(ci->ci_ssl);
-		SSL_free(ci->ci_ssl);
-		ci->ci_ssl = NULL;
-	}
-	ci->ci_tls_enabled = false;
-	ci->ci_tls_handshaking = false;
-
-	// Free the connection structure itself
-	free(ci);
-}
-
-/* Unregister a connection */
-int io_conn_unregister(int fd)
-{
-	int ret = -1;
-	struct io_context *drain_head = NULL;
-
-	pthread_mutex_lock(&conn_mutex);
-
-	int idx = fd % MAX_CONNECTIONS;
-	if (connections[idx] && connections[idx]->ci_fd == fd) {
-		TRACE("Unregistering connection fd=%d (state=%s, role=%s)", fd,
-		      io_conn_state_to_str(connections[idx]->ci_state),
-		      io_conn_role_to_str(connections[idx]->ci_role));
-
-		if (connections[idx]->ci_ssl) {
-			SSL_shutdown(connections[idx]->ci_ssl);
-			SSL_free(connections[idx]->ci_ssl);
-			connections[idx]->ci_ssl = NULL;
-			connections[idx]->ci_tls_enabled = false;
-			connections[idx]->ci_tls_handshaking = false;
-		}
-
-		/*
-		 * Drain the pending write queue before clearing ci_fd.
-		 * We must extract the list under conn_mutex but call
-		 * io_context_destroy() outside it, because destroy takes
-		 * conn_mutex internally (sequential, not nested -- safe).
-		 */
-		drain_head = connections[idx]->ci_write_pending_head;
-		connections[idx]->ci_write_pending_head = NULL;
-		connections[idx]->ci_write_pending_tail = NULL;
-		connections[idx]->ci_write_active = false;
-
-		/* Mark as unused, but keep the structure for reuse */
-		connections[idx]->ci_state = CONN_UNUSED;
-		connections[idx]->ci_fd = -1;
-		connections[idx]->ci_read_count = 0;
-		connections[idx]->ci_write_count = 0;
-		connections[idx]->ci_accept_count = 0;
-		connections[idx]->ci_connect_count = 0;
-		ret = 0;
-	} else {
-		LOG("Failed to unregister connection fd=%d - not found or mismatch",
-		    fd);
-	}
-
-	pthread_mutex_unlock(&conn_mutex);
-
-	/* Destroy drained contexts outside conn_mutex to avoid deadlock */
-	while (drain_head) {
-		struct io_context *next = drain_head->ic_write_next;
-		drain_head->ic_write_next = NULL;
-		io_context_destroy(drain_head);
-		drain_head = next;
-	}
-
-	return ret;
-}
-
-// Function to check if there are any active read operations
-bool io_conn_has_read_ops(int fd)
-{
-	bool has_ops = false;
-	pthread_mutex_lock(&conn_mutex);
-
-	int idx = fd % MAX_CONNECTIONS;
-	if (connections[idx] && connections[idx]->ci_fd == fd) {
-		has_ops = (connections[idx]->ci_read_count > 0);
-	}
-
-	pthread_mutex_unlock(&conn_mutex);
-	return has_ops;
-}
-
-// Function to check if there are any active write operations
-bool io_conn_has_write_ops(int fd)
-{
-	bool has_ops = false;
-	pthread_mutex_lock(&conn_mutex);
-
-	int idx = fd % MAX_CONNECTIONS;
-	if (connections[idx] && connections[idx]->ci_fd == fd) {
-		has_ops = (connections[idx]->ci_write_count > 0);
-	}
-
-	pthread_mutex_unlock(&conn_mutex);
-	return has_ops;
-}
-
-// Clean up connection tracking
-void io_conn_cleanup(void)
-{
-	pthread_mutex_lock(&conn_mutex);
-
-	for (int i = 0; i < MAX_CONNECTIONS; i++) {
-		if (connections[i]) {
-			if (connections[i]->ci_ssl) {
-				SSL_shutdown(connections[i]->ci_ssl);
-				SSL_free(connections[i]->ci_ssl);
-				connections[i]->ci_ssl = NULL;
-			}
-
-			free(connections[i]);
-			connections[i] = NULL;
-		}
-	}
-
-	pthread_mutex_unlock(&conn_mutex);
-}
-
-int io_socket_close(int fd, int error)
-{
-	if (fd <= 0) {
-		return -EBADF;
-	}
-
-	struct conn_info *conn = io_conn_get(fd);
-	if (conn) {
-		io_conn_set_error(fd, error);
-		io_conn_unregister(fd);
-	}
-
-	TRACE("Closing %d", fd);
-
-	io_client_fd_unregister(fd);
-	return close(fd);
-}
-
-// Utility function to get connection state as string
-const char *io_conn_state_to_str(enum conn_state state)
-{
-	switch (state) {
-	case CONN_UNUSED:
-		return "UNUSED";
-	case CONN_LISTENING:
-		return "LISTENING";
-	case CONN_ACCEPTING:
-		return "ACCEPTING";
-	case CONN_ACCEPTED:
-		return "ACCEPTED";
-	case CONN_CONNECTING:
-		return "CONNECTING";
-	case CONN_CONNECTED:
-		return "CONNECTED";
-	case CONN_READING:
-		return "READING";
-	case CONN_WRITING:
-		return "WRITING";
-	case CONN_READWRITE:
-		return "READWRITE";
-	case CONN_DISCONNECTING:
-		return "DISCONNECTING";
-	case CONN_ERROR:
-		return "ERROR";
-	default:
-		return "UNKNOWN";
-	}
-}
-
-// Helper function to convert role to string
-const char *io_conn_role_to_str(enum conn_role role)
-{
-	switch (role) {
-	case CONN_ROLE_UNKNOWN:
-		return "UNKNOWN";
-	case CONN_ROLE_CLIENT:
-		return "CLIENT";
-	case CONN_ROLE_SERVER:
-		return "SERVER";
-	case CONN_ROLE_ACCEPTED:
-		return "ACCEPTED";
-	default:
-		return "INVALID";
-	}
-}
-
-// Dump information about a specific connection
-void io_conn_dump(int fd)
-{
-	struct conn_info *ci = io_conn_get(fd);
-	if (!ci) {
-		LOG("No connection info for fd=%d", fd);
-		return;
-	}
-
-	char peer_addr[INET6_ADDRSTRLEN] = { 0 };
-	char local_addr[INET6_ADDRSTRLEN] = { 0 };
-	uint16_t peer_port = 0, local_port = 0;
-	time_t now = time(NULL);
-
-	if (ci->ci_peer_len > 0) {
-		// Cast to sockaddr_storage* instead of sockaddr*
-		addr_to_string((const struct sockaddr_storage *)&ci->ci_peer,
-			       peer_addr, INET6_ADDRSTRLEN, &peer_port);
-	}
-
-	if (ci->ci_local_len > 0) {
-		// Cast to sockaddr_storage* instead of sockaddr*
-		addr_to_string((const struct sockaddr_storage *)&ci->ci_local,
-			       local_addr, INET6_ADDRSTRLEN, &local_port);
-	}
-
-	trace_io_active_connections(ci, peer_addr, peer_port, local_addr,
-				    local_port, now, __func__, __LINE__);
-}
-
-// Dump all active connections
-void io_conn_dump_all(void)
-{
-	pthread_mutex_lock(&conn_mutex);
-
-	TRACE("=== Active Connections ===");
-	int active_count = 0;
-
-	time_t now = time(NULL);
-
-	for (int i = 0; i < MAX_CONNECTIONS; i++) {
-		if (connections[i] && connections[i]->ci_state != CONN_UNUSED) {
-			active_count++;
-			struct conn_info *ci = connections[i];
-
-			char peer_addr[INET6_ADDRSTRLEN] = { 0 };
-			char local_addr[INET6_ADDRSTRLEN] = { 0 };
-			uint16_t peer_port = 0, local_port = 0;
-
-			if (ci->ci_peer_len > 0) {
-				addr_to_string((const struct sockaddr_storage
-							*)&ci->ci_peer,
-					       peer_addr, INET6_ADDRSTRLEN,
-					       &peer_port);
-			}
-
-			if (ci->ci_local_len > 0) {
-				addr_to_string((const struct sockaddr_storage
-							*)&ci->ci_local,
-					       local_addr, INET6_ADDRSTRLEN,
-					       &local_port);
-			}
-
-			trace_io_active_connections(ci, peer_addr, peer_port,
-						    local_addr, local_port, now,
-						    __func__, __LINE__);
-		}
-	}
-
-	TRACE("Total active connections: %d", active_count);
-	TRACE("==========================");
-
-	pthread_mutex_unlock(&conn_mutex);
-}
-
-// Periodically check for timed-out connections
-int io_conn_check_timeouts(time_t timeout_seconds)
-{
-	time_t now = time(NULL);
-	int closed = 0;
-
-	pthread_mutex_lock(&conn_mutex);
-
-	for (int i = 0; i < MAX_CONNECTIONS; i++) {
-		if (connections[i] && connections[i]->ci_state != CONN_UNUSED &&
-		    connections[i]->ci_fd >= 0) {
-			// Check if connection has timed out
-			if (now - connections[i]->ci_last_activity >
-			    timeout_seconds) {
-				LOG("Connection fd=%d timed out (%ld seconds inactive)",
-				    connections[i]->ci_fd,
-				    now - connections[i]->ci_last_activity);
-
-				// Close the socket
-				close(connections[i]->ci_fd);
-
-				// Mark as unused
-				connections[i]->ci_state = CONN_UNUSED;
-				connections[i]->ci_fd = -1;
-				connections[i]->ci_read_count = 0;
-				connections[i]->ci_write_count = 0;
-				connections[i]->ci_accept_count = 0;
-				connections[i]->ci_connect_count = 0;
-
-				closed++;
-			}
-		}
-	}
-
-	pthread_mutex_unlock(&conn_mutex);
-	return closed;
-}
 
 int io_send_request(struct rpc_trans *rt)
 {
@@ -982,82 +223,4 @@ int io_handle_connect(struct io_context *ic, int result,
 
 	// Now that we're connected, prepare the RPC write request
 	return io_rpc_trans_cb(rt);
-}
-
-/*
- * io_conn_write_try_start -- atomically claim the per-fd write gate.
- *
- * If the gate is free (ci_write_active == false), set it true and return
- * true so the caller can submit a write SQE immediately.
- *
- * If the gate is held, append ic to ci_write_pending and return false.
- * The caller must return without touching ic further; io_conn_write_done()
- * will hand off the gate and call rpc_trans_writer() for the next ic.
- *
- * If fd is no longer registered we return true anyway -- the caller will
- * fail naturally on SQE submission.
- */
-bool io_conn_write_try_start(int fd, struct io_context *ic)
-{
-	bool can_start = true;
-
-	pthread_mutex_lock(&conn_mutex);
-
-	int idx = fd % MAX_CONNECTIONS;
-	if (connections[idx] && connections[idx]->ci_fd == fd) {
-		struct conn_info *ci = connections[idx];
-		if (!ci->ci_write_active) {
-			ci->ci_write_active = true;
-		} else {
-			/* Gate is held: enqueue ic for later */
-			ic->ic_write_next = NULL;
-			if (!ci->ci_write_pending_head) {
-				ci->ci_write_pending_head = ic;
-				ci->ci_write_pending_tail = ic;
-			} else {
-				ci->ci_write_pending_tail->ic_write_next = ic;
-				ci->ci_write_pending_tail = ic;
-			}
-			can_start = false;
-		}
-	}
-	/* fd gone: let caller attempt SQE and fail naturally */
-
-	pthread_mutex_unlock(&conn_mutex);
-	return can_start;
-}
-
-/*
- * io_conn_write_done -- release the per-fd write gate after a write completes.
- *
- * If there are queued writers, dequeue the head, mark it WRITE_OWNED, and
- * return it.  ci_write_active remains true so the returned ic can proceed
- * without re-claiming the gate.
- *
- * If the queue is empty, clear ci_write_active and return NULL.
- */
-struct io_context *io_conn_write_done(int fd)
-{
-	struct io_context *next_ic = NULL;
-
-	pthread_mutex_lock(&conn_mutex);
-
-	int idx = fd % MAX_CONNECTIONS;
-	if (connections[idx] && connections[idx]->ci_fd == fd) {
-		struct conn_info *ci = connections[idx];
-		if (ci->ci_write_pending_head) {
-			next_ic = ci->ci_write_pending_head;
-			ci->ci_write_pending_head = next_ic->ic_write_next;
-			next_ic->ic_write_next = NULL;
-			if (!ci->ci_write_pending_head)
-				ci->ci_write_pending_tail = NULL;
-			/* Gate stays active; transfer ownership to next_ic */
-			next_ic->ic_state |= IO_CONTEXT_WRITE_OWNED;
-		} else {
-			ci->ci_write_active = false;
-		}
-	}
-
-	pthread_mutex_unlock(&conn_mutex);
-	return next_ic;
 }
